feat(project): make GRASP use Integer
The next implementation will use Integer for the
nodes. In order to reduce the required time, the
code will be made less general.

+++ project/GRASP/src/main/java/problems/kqbf/KQBF_Inverse.java
@@ -9,22 +9,22 @@ public class KQBF_Inverse extends KQBF {
     }
 
     @Override
-    public Double evaluateQBF() {
+    public Integer evaluateQBF() {
         return -super.evaluateQBF();
     }
 
     @Override
-    public Double evaluateInsertionQBF(int i) {
+    public Integer evaluateInsertionQBF(int i) {
         return -super.evaluateInsertionQBF(i);
     }
 
     @Override
-    public Double evaluateRemovalQBF(int i) {
+    public Integer evaluateRemovalQBF(int i) {
         return -super.evaluateRemovalQBF(i);
     }
 
     @Override
-    public Double evaluateExchangeQBF(int in, int out) {
+    public Integer evaluateExchangeQBF(int in, int out) {
         return -super.evaluateExchangeQBF(in, out);
     }
